+++ src/server/__tests__/lists.test.ts
@@ -1,6 +1,6 @@
 import { Express } from "express";
 import request from "supertest";
-import { getServer } from "../../server";
+import { getServer } from "../server";
 
 describe("Lists routes", () => {
   let server: Express;
