+++ src/server/components/auth/constants.ts
@@ -1,6 +1,2 @@
 export const JWT_ALGORITHM = "HS256";
 export const JWT_EXPIRE_TIME = "5m";
-export const authRoutes = {
-  login: "/login",
-  logout: "/logout",
-};
