+++ src/server/components/auth/routes.ts
@@ -1,0 +1,4 @@
+export const authRoutes = {
+  login: "/login",
+  logout: "/logout",
+};
