+++ src/server/components/cookies/constants.ts
@@ -1,3 +1,2 @@
-export const cookiesPageRoute = "/help/cookies";
 export const ONE_YEAR = 31556952000;
 export const COOKIES_PAGE_VIEW = "cookies/cookies-page.html";
