+++ src/server/components/cookies/helpers.ts
@@ -1,0 +1,6 @@
+import { Express } from "express";
+import { cookiesRouter } from "./router";
+
+export async function initCookies(server: Express): Promise<void> {
+  server.use(cookiesRouter);
+}
