+++ src/server/components/cookies/index.ts
@@ -1,3 +1,1 @@
-export { cookiesPageRoute } from "./constants";
-export { cookiesGETController, cookiesPOSTController } from "./cookies";
-export { router as cookiesRouter } from "./router";
+export { initCookies } from "./helpers";
