+++ src/server/components/cookies/routes.ts
@@ -1,0 +1,1 @@
+export const cookiesPageRoute = "/help/cookies";
