+++ src/server/components/development/helpers.ts
@@ -1,0 +1,6 @@
+import { Express } from "express";
+import { developmentRouter } from "./router";
+
+export async function initDevelopment(server: Express): Promise<void> {
+  server.use(developmentRouter);
+}
