+++ src/server/components/development/router.ts
@@ -1,18 +1,18 @@
 /* eslint-disable @typescript-eslint/no-misused-promises */
 import express from "express";
-// import { dashboardRoutes } from "server/components/dashboard";
 import { deployDb, resetDb, promoteUser } from "./development";
+import { ensureAuthenticated } from "server/components/auth";
+import { developmentRoutes } from "./routes";
 
 export const developmentRouter = express.Router();
 
-const routes = {
-  
-}
+developmentRouter.get(`/development*`, ensureAuthenticated);
 
-developmentRouter.get(`dashboard/dev/deploy-db`, deployDb);
-
-// TODO: once prod is ready this route should be available only on localhost
-developmentRouter.get(`dashboard/dev/reset-db`, resetDb);
+// deploy db to apply future database migration changes
+developmentRouter.get(developmentRoutes.deployDb, deployDb);
 
 // TODO: once prod is ready this route should be available only on localhost
-developmentRouter.get(`dashboard/dev/promote-user`, promoteUser);
+developmentRouter.get(developmentRoutes.resetDb, resetDb);
+
+// TODO: once prod is ready this route should be available only on localhost
+developmentRouter.get(developmentRoutes.promoteUser, promoteUser);
