+++ src/server/components/development/routes.ts
@@ -1,0 +1,5 @@
+export const developmentRoutes = {
+  deployDb: "/development/deploy-db",
+  resetDb: "/development/reset-db",
+  promoteUser: "/development/promote-user",
+};
