+++ src/server/components/feedback/router.ts
@@ -1,10 +1,7 @@
 /* eslint-disable @typescript-eslint/no-misused-promises */
 import express from "express";
 import { feedbackIngest } from "./feedback";
-
-export const feedbackRoutes = {
-  postFeedback: "/feedback"
-}
+import { feedbackRoutes } from "./routes";
 
 export const feedbackRouter = express.Router();
