+++ src/server/components/feedback/routes.ts
@@ -1,0 +1,3 @@
+export const feedbackRoutes = {
+  postFeedback: "/feedback"
+};
