+++ src/server/components/healthCheck/helpers.ts
@@ -1,0 +1,6 @@
+import { Express } from "express";
+import { healthCheckRouter } from "./router";
+
+export async function initHealthCheck(server: Express): Promise<void> {
+  server.use(healthCheckRouter);
+}
