+++ src/server/components/healthCheck/routes.ts
@@ -1,0 +1,4 @@
+export const healthCheckRoutes = {
+  healthCheck: "/health-check",
+  ping: "/ping"
+};
