+++ src/server/components/lists/constants.ts
@@ -1,6 +1,6 @@
 import { countriesList } from "server/services/metadata";
 import { ServiceType } from "server/models/types";
-import { listsRoutes } from "./router";
+import { listsRoutes } from "./routes";
 
 export const DEFAULT_VIEW_PROPS = {
   countriesList,
