+++ src/server/components/lists/index.ts
@@ -1,3 +1,4 @@
-export { listsRoutes, listsRouter } from "./router";
-export { getServiceLabel, createListSearchBaseLink } from "./helpers"
+export { initLists } from "./helpers";
+export { listsRoutes } from "./routes";
+export { getServiceLabel, createListSearchBaseLink } from "./helpers";
 export * from "./types";
