+++ src/server/components/sitemap/index.ts
@@ -1,1 +1,1 @@
-export { sitemapRouter } from "./router";
+export { initSitemap } from "./helpers";
