+++ src/server/components/sitemap/routes.ts
@@ -1,0 +1,1 @@
+export const sitemapRoute = "/sitemap";
